Take advantage of @BuildSteps in the Hibernate Search extension
