+++ extensions/hibernate-search-orm-elasticsearch/deployment/src/main/java/io/quarkus/hibernate/search/orm/elasticsearch/deployment/devconsole/HibernateSearchElasticsearchDevConsoleProcessor.java
@@ -9,6 +9,7 @@ import java.util.stream.Collectors;
 
 import io.quarkus.deployment.IsDevelopment;
 import io.quarkus.deployment.annotations.BuildStep;
+import io.quarkus.deployment.annotations.BuildSteps;
 import io.quarkus.deployment.annotations.Record;
 import io.quarkus.deployment.pkg.builditem.CurateOutcomeBuildItem;
 import io.quarkus.devconsole.spi.DevConsoleRouteBuildItem;
@@ -18,9 +19,10 @@ import io.quarkus.hibernate.search.orm.elasticsearch.deployment.HibernateSearchE
 import io.quarkus.hibernate.search.orm.elasticsearch.runtime.HibernateSearchElasticsearchRuntimeConfig;
 import io.quarkus.hibernate.search.orm.elasticsearch.runtime.devconsole.HibernateSearchDevConsoleRecorder;
 
+@BuildSteps(onlyIf = HibernateSearchEnabled.class)
 public class HibernateSearchElasticsearchDevConsoleProcessor {
 
-    @BuildStep(onlyIf = { HibernateSearchEnabled.class, IsDevelopment.class })
+    @BuildStep(onlyIf = IsDevelopment.class)
     @Record(RUNTIME_INIT)
     public DevConsoleRuntimeTemplateInfoBuildItem collectBeanInfo(HibernateSearchDevConsoleRecorder recorder,
             HibernateSearchElasticsearchRuntimeConfig runtimeConfig,
@@ -33,7 +35,7 @@ public class HibernateSearchElasticsearchDevConsoleProcessor {
                 recorder.infoSupplier(runtimeConfig, persistenceUnitNames), this.getClass(), curateOutcomeBuildItem);
     }
 
-    @BuildStep(onlyIf = HibernateSearchEnabled.class)
+    @BuildStep
     @Record(value = STATIC_INIT, optional = true)
     DevConsoleRouteBuildItem invokeEndpoint(HibernateSearchDevConsoleRecorder recorder) {
         return new DevConsoleRouteBuildItem("entity-types", "POST", recorder.indexEntity());
